remove version of pbf_move_left_joystick_old that uses ticks.

+++ SerialPrograms/Source/PokemonSwSh/Programs/ShinyHuntUnattended/PokemonSwSh_CurryHunter.h
@@ -36,7 +36,7 @@ public:
 private:
     StartInGripOrGameOption START_LOCATION;
     MillisecondsOption WALK_UP_DELAY0;
-    SimpleIntegerOption<uint16_t> SMALL_POKEMON;
+    SimpleIntegerOption<uint64_t> SMALL_POKEMON;
     BooleanCheckBoxOption TAKE_VIDEO;
     SimpleIntegerOption<uint32_t> ITERATIONS;
 };
